+++ src/common.cpp
@@ -10,52 +10,75 @@
 extern Mode mode;
 
 /**
- * @brief Merge two sorted arrays into one sorted array.
+ * @brief Merge two sorted arrays into one.
  * @param result The array to merge into.
  * @param left The left array to merge.
  * @param right The right array to merge.
- * @param leftSize The size of the left array.
- * @param rightSize The size of the right array.
- * @param copy Whether to temporarily store the result in a new array.
+ * @param left_size The size of the left array.
+ * @param right_size The size of the right array.
  */
-void merge(float result[], const float left[], const float right[], const int left_size, const int right_size, const int copy) {
+void merge(float result[], const float left[], const float right[], const int left_size, const int right_size) {
     int i = 0, j = 0, k = 0;
-    float *temp;
-    if (copy)
-        temp = new float[left_size + right_size];
-    else
-        temp = result;
+
+    while (i < left_size && j < right_size) {
+        if (ACCESS(left[i]) < ACCESS(right[j]))
+            result[k++] = left[i++];
+        else
+            result[k++] = right[j++];
+    }
+
+    while (i < left_size)
+        result[k++] = left[i++];
+
+    while (j < right_size)
+        result[k++] = right[j++];
+}
+
+
+/**
+ * @brief Merge two sorted arrays into one.
+ * @param arr The array to merge into.
+ * @param left_size The size of the left array.
+ * @param right_size The size of the right array.
+ */
+void merge(float arr[], const int left_size, const int right_size) {
+    int i = 0, j = 0, k = 0;
+    float *left = arr, *right = arr + left_size;
+    const int len = left_size + right_size;
+    float *temp = new float[len];
+
     while (i < left_size && j < right_size) {
         if (ACCESS(left[i]) < ACCESS(right[j]))
             temp[k++] = left[i++];
         else
             temp[k++] = right[j++];
     }
+
     while (i < left_size)
         temp[k++] = left[i++];
+    
     while (j < right_size)
         temp[k++] = right[j++];
-    if (copy) {
-        for (int i = 0; i < left_size + right_size; i++)
-            result[i] = temp[i];
-        delete[] temp;
-    }
+    
+    for (i = 0; i < len; i++)
+        arr[i] = temp[i];
+    
+    delete[] temp;
 }
 
 
 /**
  * @brief Sort an array using merge sort.
- * @param data The array to sort.
+ * @param arr The array to sort.
  * @param size The size of the array.
- * @param result The array to store the result.
  */
 void mergeSort(float arr[], const int size) {
     if (size <= 1)
         return;
-    int mid = size / 2;
+    const int mid = size / 2;
     mergeSort(arr, mid);
     mergeSort(arr + mid, size - mid);
-    merge(arr, arr, arr + mid, mid, size - mid, 1);
+    merge(arr, mid, size - mid);
 }